create image tag for docops-* git tags
Without this, no image tags was created when a docops-* git tag was created, which meant that building the docops image failed.
This change will cause image tags of the form `docops-[0-9]+\.[0-9]+\.[0-9]+` to be created for any git tag that matches the pattern `docops-[0-9]+\.[0-9]+\.[0-9]+`. Since the git tag has to contain `docops`, other images should not be affected.